feat: library cleanup/deletion, upload UI; retire the passive image cache
Library management
- `clean_library` only ever deleted media whose status was already not-Ready, so a
  file removed outside Longbox left a permanently un-removable ghost: three books
  on this deployment sat at status READY while their directory no longer existed
  on disk. Clean now also verifies that READY files are still present.
  Critically, it first checks the library root is a readable NON-EMPTY directory:
  an unmounted share usually leaves its mount point behind as an empty dir, so a
  plain exists() check would pass and every book under it would look deleted.
  When the root fails that check the sweep is skipped entirely and reported,
  rather than deleting the library's contents. Only a definitive NotFound counts
  as missing -- permission denied, EIO and stale NFS handles are indeterminate.
- The "Clean library" button was disabled whenever the missing-entities table was
  empty, and that table only lists records already flagged MISSING. In exactly the
  reported scenario -- READY ghosts, nothing flagged -- the fix was unreachable.
- `delete_media` was a soft delete that never touched the filesystem, so there was
  no way to remove a book and its file. Adds `deleteMediaPermanently`, guarded,
  with the path canonicalized and verified to be contained within its library
  before anything is removed; the canonical path is what gets deleted, not the
  stored one. File is removed before the row: if the unlink fails we abort with
  the DB still describing disk truthfully, and if the row delete fails the result
  is a ghost -- which the clean sweep above now repairs.

Upload
- Book and series upload mutations already existed but had no UI. Adds an upload
  section to library settings using the existing directory picker, with real
  progress, server error text surfaced verbatim, and a success state explaining
  that a library scan has been queued.
- Book/series uploads enforced NO size limit at all, so `max_file_upload_size`
  bounded nothing; meanwhile the thumbnail mutations were bounded by it rather
  than by `max_image_upload_size`. Both corrected, so raising the file limit no
  longer also permits enormous thumbnails.
- `maxFileUploadSize` was typed `Int` while holding a `usize`. Values above ~2.1GB
  serialized fine but the SDL was lying, and a spec-validating client would reject
  them. Now `Float`, exact to 2^53.

Offline
- Retires the passive read-through cache. Covers are ~50KB server-side now and
  image responses carry ETag + stale-while-revalidate, so the browser's own HTTP
  cache does this correctly; the passive layer duplicated it, shared a storage
  bucket with explicit downloads, and pinned regenerated thumbnails for a day.
  Explicit "download for offline" is untouched -- it is the actual offline
  reading feature and was verified end-to-end with the network cut.
- The v3->v4 migration purges the retired cache's blobs, which shared a bucket
  with downloads. Ownership is taken only from DownloadRecords and URLs are
  matched in both stored and resolved-absolute form, so an unrecognized shape
  leaks a passive blob rather than destroying a downloaded book.

Tests: 513 -> 449 frontend. Removed the passive-cache suites, deleted
`blobStore.estimateUsage` (no production callers -- it existed for that cache),
and folded an overlapping preload case into the stronger one.

+++ crates/graphql/src/query/config.rs
@@ -4,10 +4,22 @@ use models::shared::enums::UserPermission;
 
 use crate::{data::CoreContext, guard::PermissionGuard};
 
+/// The upload-related slice of the server configuration. This exists so clients can
+/// render an accurate upload control -- whether uploads are permitted at all, and how
+/// large a file may be -- instead of discovering both by way of a failed mutation.
 #[derive(Default, SimpleObject)]
 pub struct UploadConfig {
+	/// Whether `enable_upload` is set on the server. When false, every upload mutation
+	/// is rejected by [`crate::guard::OptionalFeatureGuard`].
 	enabled: bool,
-	max_file_upload_size: usize,
+	/// The maximum size, in bytes, of a single uploaded file.
+	///
+	/// Deliberately a GraphQL `Float` (f64) rather than an `Int`. `Int` is a *signed
+	/// 32-bit* value per the GraphQL spec, so it tops out at ~2.1 GB, and deployments
+	/// routinely configure a larger ceiling (5 GB on ours). `f64` represents every
+	/// integer up to 2^53 exactly, so a byte count is lossless well past any plausible
+	/// upload limit, and it maps to a plain `number` in the generated TS client.
+	max_file_upload_size: f64,
 }
 
 #[derive(Default)]
@@ -27,7 +39,7 @@ impl ConfigQuery {
 		let config = ctx.data::<CoreContext>()?.config.as_ref();
 		Ok(UploadConfig {
 			enabled: config.enable_upload,
-			max_file_upload_size: config.max_file_upload_size,
+			max_file_upload_size: config.max_file_upload_size as f64,
 		})
 	}
 }
